+++ pepper/rc/changelog.md
@@ -18,6 +18,7 @@
 - added support for recursive variable expansions (that is, it's possible to have a `@var(...)` inside a `@var(...)`)
 - changed parsing of expansion arguments to be more robust like command parsing
 - added `@command-source()` expansion which expands into a macro command's source
+- changed `remedybg-plugin` to support newer events from version `0.3.8.6`
 
 # 0.30.0
 - added `insert-text` command